refactor(adapters): ApiAdapter を 8 ドメインに分割 (#707)
interface (types.ts) と Misskey 実装 (904 行の MisskeyApi クラス) を
notes / users / collections / chat / notifications / drive / charts /
server の 8 ドメインに分割。呼び出し側はフラットな
`adapter.api.getNote(...)` のまま — ApiAdapter はドメイン interface の
合成、実装は createMisskeyApi() がドメインファクトリを合成して返す。

- MisskeyApi クラス → createMisskeyApi ファクトリ (利用 2 箇所を更新)
- 共有状態は MisskeyApiContext (accountId / hasToken / requireAuth) に集約
- dead code を削除: createChatMessage / getCachedChat* 3 メソッドは
  adapter 経由の呼び出し元ゼロ (chat の書込・DB キャッシュは Misskey
  専用機能の例外として commands.* 直呼びに統一済みだった)
